refactor: ipc logic
