+++ zellij-client/src/web_client/mod.rs
@@ -3,6 +3,7 @@ mod control_message;
 mod authentication;
 mod connection_manager;
 mod http_handlers;
+mod ipc_listener;
 mod message_handlers;
 mod server_listener;
 mod session_management;
@@ -31,19 +32,15 @@ use nix::sys::stat::{umask, Mode};
 
 use interprocess::unnamed_pipe::pipe;
 use std::io::{prelude::*, BufRead, BufReader};
-use tokio::io::{AsyncBufReadExt, AsyncReadExt};
-use tokio::net::{UnixListener, UnixStream};
 use tokio::runtime::Runtime;
 use tower_http::cors::CorsLayer;
-use uuid::Uuid;
-use zellij_utils::consts::WEBSERVER_SOCKET_PATH;
 use zellij_utils::input::{config::Config, options::Options};
-use zellij_utils::ipc::InstructionForWebServer;
 
 use authentication::auth_middleware;
 use http_handlers::{
     create_new_client, get_static_asset, login_handler, serve_html, version_handler,
 };
+use ipc_listener::listen_to_web_server_instructions;
 use types::{
     AppState, ClientOsApiFactory, ConnectionTable, RealClientOsApiFactory, RealSessionManager,
     SessionManager,
@@ -147,55 +144,6 @@ pub fn start_web_client(config: Config, config_options: Options, run_daemonized:
         None,
         None,
     ));
-}
-
-pub async fn create_webserver_receiver(
-) -> Result<UnixStream, Box<dyn std::error::Error + Send + Sync>> {
-    let id = Uuid::new_v4();
-    let socket_path = WEBSERVER_SOCKET_PATH.join(format!("{}", id));
-
-    if socket_path.exists() {
-        tokio::fs::remove_file(&socket_path).await?;
-    }
-
-    let listener = UnixListener::bind(&socket_path)?;
-    let (stream, _) = listener.accept().await?;
-    Ok(stream)
-}
-
-pub async fn receive_webserver_instruction(
-    receiver: &mut UnixStream,
-) -> std::io::Result<InstructionForWebServer> {
-    let mut buffer = Vec::new();
-    receiver.read_to_end(&mut buffer).await?;
-    let cursor = std::io::Cursor::new(buffer);
-    rmp_serde::decode::from_read(cursor)
-        .map_err(|e| std::io::Error::new(std::io::ErrorKind::Other, e))
-}
-
-// here we listen to internal web server instructions coming over an IPC channel
-async fn listen_to_web_server_instructions(server_handle: Handle) {
-    loop {
-        match create_webserver_receiver().await {
-            Ok(mut receiver) => loop {
-                match receive_webserver_instruction(&mut receiver).await {
-                    Ok(instruction) => match instruction {
-                        InstructionForWebServer::ShutdownWebServer => {
-                            server_handle.shutdown();
-                            break;
-                        },
-                    },
-                    Err(e) => {
-                        log::error!("Failed to process web server instruction: {}", e);
-                        break;
-                    },
-                }
-            },
-            Err(e) => {
-                log::error!("Failed to create receiver: {}", e);
-            },
-        }
-    }
 }
 
 pub async fn serve_web_client(
